Minor fixes and comments
