+++ src/ast.hpp
@@ -9,11 +9,20 @@
 #include "lex.hpp"
 
 namespace Ast::Nodes {
+    /// @brief Base class for all AST nodes
     class NodeBase {
     public:
         virtual ~NodeBase() = default;
+        /// @brief Returns a string representation of the node
+        /// @return A string representation of the node
         virtual std::string show() = 0;
+        /// @brief Generates LLVM IR for the node
+        /// @param context The context to generate the IR in
+        /// @return The generated LLVM value
         virtual llvm::Value* generate(Context& context) = 0;
+        /// @brief Parses a node from the lexer stream
+        /// @param s The lexer stream to parse from
+        /// @return A unique pointer to the parsed node
         static std::unique_ptr<NodeBase> parse(Lexer::Stream& s);
     };
 
@@ -23,15 +32,25 @@ namespace Ast::Nodes {
 
         static std::unique_ptr<Expr> parse(Lexer::Stream& s);
 
+        /// @brief Parses an expression from the lexer stream
+        /// @param s The lexer stream to parse from
+        /// @return A unique pointer to the parsed expression
         static std::unique_ptr<Expr> parseExpr(Lexer::Stream& s);
+        /// @brief Parses a term from the lexer stream
+        /// @param s The lexer stream to parse from
+        /// @return A unique pointer to the parsed term
         static std::unique_ptr<Expr> parseTerm(Lexer::Stream& s);
     };
 
+    /// @brief Smart pointer type for expressions
     using pExpr = std::unique_ptr<Expr>;
 
+    /// @brief Expression node for binary operations
     class ExprOp : public Expr {
     public:
+        /// @brief Type of binary operation
         enum Type { Add, Sub, Mul, Div } type;
+        /// @brief operands
         std::unique_ptr<Expr> left, right;
 
         ExprOp(Type type, pExpr left, pExpr right)
@@ -42,8 +61,10 @@ namespace Ast::Nodes {
         llvm::Value* generate(Context& ctx) override;
     };
 
+    /// @brief Expression node for integer literals
     class ExprLitInt : public Expr {
     private:
+        /// @brief The integer value
         uint64_t val;
         
     public:
@@ -54,7 +75,10 @@ namespace Ast::Nodes {
         virtual llvm::Value* generate(Context& ctx) override;
     };
 
+    /// @brief Expression node for floating-point literals
     class ExprLitFloat : public Expr {
+    private:
+        /// @brief The floating-point value
         double val;
 
     public:
@@ -65,8 +89,10 @@ namespace Ast::Nodes {
         llvm::Value* generate(Context& ctx) override;
     };
 
+    /// @brief Expression node for variable references
     class ExprVar : public Expr {
     private:
+        /// @brief The variable name
         const std::string name;
 
     public:
@@ -79,9 +105,12 @@ namespace Ast::Nodes {
         llvm::Value* generate(Context& ctx) override;
     };
 
+    /// @brief Expression node for variable assignments
     class ExprAssign : public Expr {
     private:
+        /// @brief The variable name
         const std::string name;
+        /// @brief The expression to assign to the variable
         pExpr expr;
 
     public:
@@ -94,9 +123,10 @@ namespace Ast::Nodes {
         llvm::Value* generate(Context& ctx) override;
     };
 
-    // Rust-style block (returns a value)
+    /// @brief Expression node for conditional statements
     class ExprBlock : public Expr {
     private:
+        /// @brief The list of nodes in the block
         std::vector<std::unique_ptr<NodeBase>> nodes;
     
     public:
@@ -111,13 +141,16 @@ namespace Ast::Nodes {
         llvm::Value* generate(Context& ctx) override;
     };
 
+    /// @brief Node for if statements
     class If : public Expr {
     private:
+        /// @brief The condition expression
         pExpr cond;
-        pExpr expr;
+        /// @brief The expression to execute if the condition is true
+        std::unique_ptr<NodeBase> expr;
 
     public:
-        If(pExpr expr, pExpr cond)
+        If(std::unique_ptr<NodeBase> expr, pExpr cond)
         : expr(std::move(expr)), cond(std::move(cond)) { }
 
         static std::unique_ptr<If> parse(Lexer::Stream& s);
@@ -126,9 +159,12 @@ namespace Ast::Nodes {
         llvm::Value* generate(Context& ctx) override;
     };
 
+    /// @brief Node for variable declarations
     class Let : public NodeBase {
     private:
+        /// @brief The target variable name
         std::string target;
+        /// @brief The expression for the variable's initial value
         pExpr expr;
 
     public:
@@ -142,9 +178,10 @@ namespace Ast::Nodes {
         llvm::Value* generate(Context& ctx) override;
     };
 
+    /// @brief Node for return statements
     class Return : public NodeBase {
     private:
-        // std::shared_ptr<llvm::Function*> parent_fn; // TODO
+        /// @brief The expression to return
         pExpr expr;
     
     public:
@@ -158,10 +195,13 @@ namespace Ast::Nodes {
         llvm::Value* generate(Context& ctx) override;
     };
 
+    /// @brief Node for function definitions
     class Function : public NodeBase {
     private:
+        /// @brief The function name
         std::string name;
         // TODO: Add args
+        /// @brief The function body block
         std::unique_ptr<NodeBase> block;
 
     public:
@@ -178,8 +218,10 @@ namespace Ast::Nodes {
 };
 
 namespace Ast {
+    /// @brief The root AST node representing the entire program
     class Program {
     private:
+        /// @brief The list of nodes in the program
         std::vector<std::unique_ptr<Nodes::NodeBase>> content;
        
     public:
