+++ src/lex.hpp
@@ -12,9 +12,10 @@ namespace Ast::Nodes {
     class NodeBase;
 }
 
+/// @brief Lexer namespace
 namespace Lexer {
 
-// finite state machine
+/// @brief Character types used during lexing
 enum class CharType {
     Invalid,
     Alpha,
@@ -26,8 +27,12 @@ enum class CharType {
     Brace,
 };
 
+/// @brief Get the character type of a given character
+/// @param c Character to classify
+/// @return Character type
 CharType getCharType(char c);
 
+/// @brief Token types used in the lexer
 enum class Type {
     Invalid,
 
@@ -43,31 +48,55 @@ enum class Type {
     BraceClose,
 };
 
+/// @brief Classify a token based on its content and ending character type
+/// @param content Content of the token
+/// @param state Ending character type of the token
+/// @return Token type
 Type classify(std::string& content, CharType state);
 
+/// @brief Token representation
 class Token {
 public:
+    /// @brief Construct a new Token object
     std::string content;
+    /// @brief Line number where the token is located
     uint32_t line;
+    /// @brief Column number where the token starts
     Type type;
 };
 
+/// @brief Stream of tokens produced by the lexer
 class Stream {
 private:
+    /// @brief Current index in the token stream
     uint32_t index = 0;
 
 public:
+    /// @brief List of tokens in the stream
     std::vector<Token> content;
 
+    /// @brief Check if there are more tokens in the stream
     bool has();
+    /// @brief Peek at the next token without consuming it
+    /// @return Pointer to the next token
     std::unique_ptr<Token> peek();
+    /// @brief Pop the next token from the stream
+    /// @return Pointer to the popped token
     std::unique_ptr<Token> pop();
+    /// @brief Move back one token in the stream
     void back() { index--; }
+    /// @brief Expect the next token to match a specific content
+    /// @param should Expected token content
+    /// @param location Source location for error reporting
     void expect(
         const char* should, 
         const std::source_location& location = std::source_location::current()
     );
 
+    /// @brief Expect the next token to match a specific content and parse a node
+    /// @param should Expected token content
+    /// @param nodes_parsed Parsed node to be filled
+    /// @param location Source location for error reporting
     void expect(
         const char* should,
         std::unique_ptr<Ast::Nodes::NodeBase>& nodes_parsed,
@@ -75,6 +104,9 @@ public:
     );
 };
 
+/// @brief Tokenize a source file into a token stream
+/// @param source_path Path to the source file
+/// @return Token stream
 Stream tokenize(std::string& source_path);
 
 }
