+++ src/main.cpp
@@ -51,9 +51,9 @@ int main(int argc, char** argv) {
     auto root = Ast::Program::parse(tokens);
     std::cout << "done\n";
 
-    std::cout << "--- parsed source listing ---\n";
-    root.show(std::cout);
-    std::cout << '\n';
+    // std::cout << "--- parsed source listing ---\n";
+    // root.show(std::cout);
+    // std::cout << '\n';
 
     Context ctx;
