+++ src/syscall.hpp
@@ -5,4 +5,7 @@
 #include <llvm/IR/Function.h>
 #include <llvm/IR/LLVMContext.h>
 
+/// @brief Generates an inline assembly instruction for a syscall to write to stdout
+/// @param ctx The LLVM context
+/// @return The inline assembly instruction
 llvm::InlineAsm* syscall_exit(llvm::LLVMContext& ctx);
